backend/NeuroSense-Backend/pyproject.toml
[project]
name = "backend-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.2",

    "flask-pymongo>=3.0.1",
    "joblib>=1.5.3",
    "pandas>=2.3.3",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.6.1",

]

